refactor: standardise apps
