+++ apps/example-reducers/src/preload/index.ts
@@ -4,14 +4,14 @@ import { preloadZustandBridge } from 'zutron/preload';
 import 'wdio-electron-service/preload';
 import type { Handlers } from 'zutron';
 
-import type { State } from '../features/index.js';
+import type { AppState } from '../features/index.js';
 
-export const { handlers } = preloadZustandBridge<State>();
+export const { handlers } = preloadZustandBridge<AppState>();
 
 contextBridge.exposeInMainWorld('zutron', handlers);
 
 declare global {
   interface Window {
-    zutron: Handlers<State>;
+    zutron: Handlers<AppState>;
   }
 }
