+++ apps/example-separate-handlers/src/features/counter/index.ts
@@ -1,4 +1,4 @@
-import { type Store } from '..';
+import type { Store } from '../../main/store.js';
 
 export const handlers = (store: Store) => ({
   'COUNTER:INCREMENT': () => store.setState((state) => ({ counter: state.counter + 1 })),
