+++ apps/example-separate-handlers/src/features/index.ts
@@ -1,16 +1,10 @@
 import { handlers as counterHandlers } from './counter/index.js';
+import type { Store } from '../main/store.js';
 
-export const actionHandlers = (store: Store, initialState: State) => ({
+export type Handlers = Record<string, () => void>;
+export type AppState = { counter: number };
+
+export const actionHandlers = (store: Store, initialState: AppState) => ({
   ...counterHandlers(store),
   'STORE:RESET': () => store.setState(() => initialState),
 });
-
-export type Subscribe = (listener: (state: State, prevState: State) => void) => () => void;
-export type Handlers = Record<string, () => void>;
-export type State = { counter: number };
-export type Store = {
-  getState: () => State;
-  getInitialState: () => State;
-  setState: (stateSetter: (state: State) => State) => void;
-  subscribe: Subscribe;
-};
