+++ apps/example-separate-handlers/src/main/store.ts
@@ -1,9 +1,18 @@
 import { createStore } from 'zustand/vanilla';
 
-import type { State } from '../features/index.js';
+import type { AppState } from '../features/index.js';
 
-export const initialState = {
+type Subscribe = (listener: (state: AppState, prevState: AppState) => void) => () => void;
+
+export type Store = {
+  getState: () => AppState;
+  getInitialState: () => AppState;
+  setState: (stateSetter: (state: AppState) => AppState) => void;
+  subscribe: Subscribe;
+};
+
+export const initialState: AppState = {
   counter: 0,
 };
 
-export const store = createStore<State>()(() => initialState);
+export const store = createStore<AppState>()(() => initialState);
