+++ apps/example-separate-handlers/src/renderer/hooks/useStore.ts
@@ -1,4 +1,5 @@
 import { createUseStore } from 'zutron';
-import type { State } from '../../features/index.js';
 
-export const useStore = createUseStore<State>(window.zutron);
+import type { AppState } from '../../features/index.js';
+
+export const useStore = createUseStore<AppState>(window.zutron);
