+++ apps/example-store-handlers/src/features/counter/index.ts
@@ -1,4 +1,4 @@
-import type { Store } from '../index.js';
+import type { Store } from '../../main/store.js';
 
 export const handlers = (setState: Store['setState']) => ({
   'COUNTER:INCREMENT': () => setState((state) => ({ counter: state.counter + 1 })),
